Tweak printFunction
