+++ src/language-js/print/function.js
@@ -41,6 +41,8 @@ const {
   getComments,
   CommentCheckFlags,
   isCallLikeExpression,
+  isCallExpression,
+  getCallArguments,
 } = require("../utils");
 const { locEnd } = require("../loc");
 const {
@@ -50,8 +52,22 @@ const {
 const { printPropertyKey } = require("./property");
 const { printFunctionTypeParameters } = require("./misc");
 
-function printFunctionDeclaration(path, print, options, expandArg) {
+function printFunction(path, print, options, args) {
   const node = path.getValue();
+
+  let expandArg = false;
+  if (
+    (node.type === "FunctionDeclaration" ||
+      node.type === "FunctionExpression") &&
+    args &&
+    args.expandLastArg
+  ) {
+    const parent = path.getParentNode();
+    if (isCallExpression(parent) && getCallArguments(parent).length > 1) {
+      expandArg = true;
+    }
+  }
+
   const parts = [];
 
   // For TypeScript the TSDeclareFunction node shares the AST
@@ -260,7 +276,7 @@ function printArrowChain(
   ]);
 }
 
-function printArrowFunctionExpression(path, options, print, args) {
+function printArrowFunction(path, options, print, args) {
   let node = path.getValue();
   /** @type {Doc[]} */
   const signatures = [];
@@ -427,7 +443,7 @@ function printReturnType(path, print, options) {
 }
 
 // `ReturnStatement` and `ThrowStatement`
-function printReturnAndThrowArgument(path, options, print) {
+function printReturnOrThrowArgument(path, options, print) {
   const node = path.getValue();
   const semi = options.semi ? ";" : "";
   const parts = [];
@@ -475,16 +491,16 @@ function printReturnAndThrowArgument(path, options, print) {
 }
 
 function printReturnStatement(path, options, print) {
-  return ["return", printReturnAndThrowArgument(path, options, print)];
+  return ["return", printReturnOrThrowArgument(path, options, print)];
 }
 
 function printThrowStatement(path, options, print) {
-  return ["throw", printReturnAndThrowArgument(path, options, print)];
+  return ["throw", printReturnOrThrowArgument(path, options, print)];
 }
 
 module.exports = {
-  printFunctionDeclaration,
-  printArrowFunctionExpression,
+  printFunction,
+  printArrowFunction,
   printMethod,
   printReturnStatement,
   printThrowStatement,
